libs/bpm-ai-connectors-c8/pyproject.toml
[tool.poetry]
name = "bpm-ai-connectors-c8"
version = "1.0.0"
description = "AI connectors for Camunda 8 based on bpm-ai."
authors = ["Bennet Krause <bennet.krause@holisticon.de>"]
repository = "https://github.com/holunda-io/bpm-ai"
homepage = "https://www.holisticon.de/"
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyzeebe = "^3.0.4"
bpm-ai = "1.0.2"
pydantic = "^2.5.3"
pytest-xprocess = "^0.23.0"


[tool.poetry.group.dev.dependencies]
setuptools = "^68.2.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^0.23.3"
pytest-zeebe = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
